Switch from millidegrees to degrees for Temperature display Thanks to @Photon89

+++ common/graph-data.h
@@ -101,8 +101,8 @@ typedef struct _DiskData {
 
 typedef struct _TemperatureData {
 	gchar name[30];
-	guint value;
-	guint max;
+	double value;
+	double max;
 	AutoScaler scaler;
 } TemperatureData;
